WIP: added new routes component
